+++ Win2MQTT.sln
@@ -12,8 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		.gitattributes = .gitattributes
 		.gitignore = .gitignore
-		hass-sample.yaml = hass-sample.yaml
 		README.md = README.md
+		hass-example.yaml = hass-example.yaml
 		Roadmap.md = Roadmap.md
 	EndProjectSection
 EndProject
