+++ src/Win2Mqtt.Infra/SensorDataCollector.cs
@@ -24,26 +24,26 @@ namespace Win2Mqtt.Client.Mqtt
                 foreach (var drive in Drives.GetDriveStatus())
                 {
                     var topic = "drive/" + drive.DriveName;
-                    data.Add($"{topic}/totalsize", drive.TotalSize.ToString(CultureInfo.CurrentCulture));
-                    data.Add($"{topic}/availablefreespace", drive.AvailableFreeSpace.ToString(CultureInfo.CurrentCulture));
+                    data.Add($"{topic}/sizetotal", drive.TotalSize.ToString(CultureInfo.CurrentCulture));
+                    data.Add($"{topic}/sizefree", drive.AvailableFreeSpace.ToString(CultureInfo.CurrentCulture));
                     data.Add($"{topic}/percentfree", drive.PercentFree.ToString(CultureInfo.CurrentCulture));
                 }
             }
             if (_options.Sensors.FreeMemorySensor)
             {
-                data.Add("freememory", string.Format("{0}MB", Memory.GetFreeMemory()));
-            }
-            if (_options.Sensors.NetworkSensor)
-            {
-                data.Add("networkavailable", NetworkStatus.IsNetworkAvailable() ? "1" : "0");
+                data.Add("freememory", Memory.GetFreeMemory().ToString(CultureInfo.CurrentCulture));
             }
             if (_options.Sensors.CpuSensor)
             {
-                data.Add("cpuprocessortime", string.Format("{0}%", Processor.GetCpuProcessorTime()));
+                data.Add("cpuprocessortime", Processor.GetCpuProcessorTime().ToString(CultureInfo.CurrentCulture));
             }
             if (_options.Sensors.IsComputerUsed)
             { 
-                data.Add("binary_sensor/inuse", Processor.GetIdleTime().TotalSeconds <= 30 ? "1" : "0");
+                data.Add("binary_sensor/inuse", (Processor.GetIdleTime().TotalSeconds <= 30).ToMqttBoolean());
+            }
+            if (_options.Sensors.NetworkSensor)
+            {
+                data.Add("binary_sensor/network_available", NetworkStatus.IsNetworkAvailable().ToMqttBoolean());
             }
             return Task.FromResult<IDictionary<string, string>>(data);
         }
