+++ src/Win2Mqtt.Service/Win2Mqtt.Service.csproj
@@ -5,11 +5,16 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <UserSecretsId>dotnet-Win2Mqtt.Service-1b9a5605-98fc-4784-bb9d-1b6a3ef40c18</UserSecretsId>
-	  <SupportedOSPlatformVersion>10.0.18362.0</SupportedOSPlatformVersion>
+	<SupportedOSPlatformVersion>10.0.18362.0</SupportedOSPlatformVersion>
+	  <OutputType>exe</OutputType>
+	  <PublishSingleFile Condition="'$(Configuration)' == 'Release'">true</PublishSingleFile>
+	  <RuntimeIdentifier>win-x64</RuntimeIdentifier>
+	  <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+	<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Options.DataAnnotations" Version="8.0.0" />
     <PackageReference Include="Serilog" Version="3.1.1" />
     <PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
